# Optimize NEI rendering