I'm not sure if it's universal, but NEI is notoriously slow on my computer. This PR contains some trivial optimizations to make it run quite a bit faster. Without them, my angelica dev env idles between 50 to 60 fps. With them, my dev env idles around 130 to 140 fps.

The first one is to redirect `glGetInteger(GL_MODELVIEW_STACK_DEPTH)` to use the java matrix stack instead of deferring to the native method (along with `GL_PROJECTION_STACK_DEPTH` since I was in there, but that isn't needed here). This change shouldn't have any effect since the matrix stack already mirrors the gl matrix stack.

Also, I removed an Iterator allocation in a hot path (the `pushState`/`popState` methods). I swapped out the `Set` to a `List` so that I could use an index instead of the iterator. This is a relatively minor improvement, but it was a fairly safe one so I did it anyways.

To take these profiles, I uncapped the frame rate and turned off vsync & the alternative frame limiter. I also shrunk the render distance to 2 chunks (the minimum). I opened NEI so that it was rendering, and let my game sit for around a minute. Note that the total time taken is around the same since they're both at 100% utilization due to the uncapped fps (the main thread, not cpu).

Without:
https://spark.lucko.me/TT1Clyfsa5
![image](https://github.com/user-attachments/assets/94cd3019-6b68-431a-8982-8fbc7bd2eeda)
![image](https://github.com/user-attachments/assets/6795f7b5-8247-4ed6-80a6-55c7289f60fa)


With:
https://spark.lucko.me/FGG1fIRT2u
![image](https://github.com/user-attachments/assets/78b7c93f-ce2b-46ab-9af8-e3d316ed6562)
![Screenshot From 2025-02-22 18-03-00](https://github.com/user-attachments/assets/9087e57b-d122-433a-bc5f-20c2d2b70f35)